# Overview
Simplification and optimization of WebGL rendering pipeline. Mainly 3 objects were the focus of the otpimizations: the `flush` and `drawImage` methods and the `ContextStateStack` class in `WebGLContext2D`.

# Changes
- The `drawImage` method used to include several state checks on the image and WebGL context, they are now gone as they seem unnecessary. If they happened to be required, they should be added at a different place where they would execute only once and not every single frame.
- The `flush` method and associated methods in `Shader` are now doing program switching more efficiently.
- Textures are now wrapped into `TextureWrapper` objects that include a reference on the original image. The goal was to avoid unoptimizing images by sticking properties onto them.
- `ContextStateStack` has been removed as state properties of the context are now directly attached to the context, reducing the chain of property accesses.
- All the checks on the existence of the `gl` property were removed. WebGL context is initialized when the `WebGLContext2D` instance is created and there is no need to check if it was properly initialized in every method accessing the WebGL API.

# Performance
Overall there seems to be an obvious performance gain in terms of CPU usage.
As usual, the relative performance gain (5~20%) depends on the test, the device and measurement method.

## Chrome - MacBook pro
### Everwing - start menu - chrome profile
#### Before
<img width="400" alt="everwing-startmenu-heavy" src="https://user-images.githubusercontent.com/1897225/30098106-86bd0268-931b-11e7-8476-3e5237cd92af.png">

#### After
<img width="414" alt="everwing-startmenu-heavy" src="https://user-images.githubusercontent.com/1897225/30098090-76af065a-931b-11e7-89fa-4293d765f841.png">
The `drawImage`  and `flush` methods are now faster, The `set` method has disappeared. also the % used by the program is remain roughly identical, meaning that the performance boost is reflected in the program usage.

### Everwing - gameplay - chrome profile
#### Before
<img width="410" alt="everwing-gameplay-heavy" src="https://user-images.githubusercontent.com/1897225/30098159-c7dfeef4-931b-11e7-8fb0-bb6d5a17f628.png">

#### After
<img width="434" alt="everwing-gameplay-heavy" src="https://user-images.githubusercontent.com/1897225/30098177-d1e7a7a2-931b-11e7-9473-7e2a619fa015.png">
Here as well the `drawImage`, `flush` and `set` methods are optimized. % of program remain identical.

### Everwing - gameplay - chrome profile compilation
The following screenshots show the average non-idle CPU time profiled by chrome across 6 gameplay runs > 10s:
#### Before
<img width="154" alt="everwing-gameplay" src="https://user-images.githubusercontent.com/1897225/30098331-75d130a4-931c-11e7-9f56-2e5d94cbec00.png">

#### After
<img width="157" alt="everwing-gameplay" src="https://user-images.githubusercontent.com/1897225/30098327-72ebd948-931c-11e7-91ba-e8923b29f3d8.png">

On this benchmark the gain appears consequent: (13.8 - 11) / 13.8 = 20% boost!

### Everwing - gameplay - duration of engine tick method on chrome
#### Before
<img width="275" alt="everwing-startmenu-engine-tick" src="https://user-images.githubusercontent.com/1897225/30099697-778ad86e-9321-11e7-972a-9b132ae38cf6.png">

#### After
<img width="270" alt="everwing-startmenu-engine-tick" src="https://user-images.githubusercontent.com/1897225/30099701-7c9a16d0-9321-11e7-9976-3673d6921754.png">

Here as well the performance gain is obvious: (0.66 - 0.56) / 0.66 = 15%

## Chrome - iPhone 7
### Everwing - start menu - safari profile
#### Before
<img width="490" alt="everwing-startmenu-iphone7" src="https://user-images.githubusercontent.com/1897225/30100180-effe6170-9322-11e7-8965-d29ce7ff2f63.png">

#### After
<img width="453" alt="everwing-startmenu-iphone7" src="https://user-images.githubusercontent.com/1897225/30100193-f8ba4590-9322-11e7-91fc-33438e3811cd.png">

Just as an chrome, the `drawImage` and `set` methods have been optimized. For some reason the `flush` method never appeared (self time too insignificant). Overall benefit is not obvious on this profile as the idle time usually is hidden under the `drawElements` self time.
The optimized was possibly slow to kick in? It would seem so as subsequent profiles played in the favor of the changes.

### Everwing - start menu - safari profile
#### Before
<img width="456" alt="everwing-gameplay-iphone7" src="https://user-images.githubusercontent.com/1897225/30100989-afa4576c-9325-11e7-9629-8e25372b7dc9.png">

#### After
<img width="457" alt="everwing-gameplay-iphone7" src="https://user-images.githubusercontent.com/1897225/30100299-54bf4106-9323-11e7-8b98-2da1832a2ffc.png">

On this profile the performance boost is quite obvious (optimizer possibly played in favor of the changes) if we consider the idle time as being hidden in the `drawElements` self time: (32.6 - 24.1) / 32.6 = 26% boost

# Conclusion
Overall I think this work will provide a 10~15% cpu boost (even though some profiles seem to indicate a higher gain). Analytics will tell us but slowest devices should now run Everwing at 55fps. Meaning that we probably are 2 optimizations away from providing an average of 60fps on the slowest devices on Everwing (up from 30fps 3 months ago).